amd/common/block/spi/fch_spi_ctrl: use uintptr_t for addresses
This fixes a build failure when trying to build the code in 64 bit mode.

TEST=Timeless build for Chausie results in identical image.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/62052
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/amd/common/block/spi/fch_spi_ctrl.c
@@ -38,7 +38,7 @@ enum spi_dump_state_phase {
 static void dump_state(enum spi_dump_state_phase phase)
 {
 	u8 dump_size;
-	u32 addr;
+	uintptr_t addr;
 
 	if (!CONFIG(SOC_AMD_COMMON_BLOCK_SPI_DEBUG))
 		return;
